requirements.txt
flask==3.0.2
numpy==1.24.3
tensorflow==2.15.0
opencv-python==4.9.0.80
pillow==10.2.0
scikit-learn==1.4.1.post1
matplotlib==3.8.3